requirements.txt
# === HSC-Edu Dependencies ===

# --- PDF Extraction ---
pymupdf>=1.25.0
pdfplumber>=0.11.0

# --- OCR ---
paddleocr>=2.9.0
paddlepaddle>=3.0.0
easyocr>=1.7.0

# --- Table Extraction ---
camelot-py[cv]>=0.11.0

# --- LLM & Embeddings ---
google-genai>=1.0.0
openai>=1.60.0
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-community>=0.3.0

# --- Vector Store ---
qdrant-client>=1.12.0
chromadb>=0.6.0

# --- Metadata Store ---
pymongo>=4.9.0

# --- Data Processing ---
pandas>=2.2.0
numpy>=2.0.0
pyyaml>=6.0.0
pydantic>=2.10.0

# --- Utilities ---
python-dotenv>=1.0.0
tqdm>=4.67.0
tiktoken>=0.8.0

# --- API Backend (Phase 4) ---
fastapi>=0.115.0
uvicorn>=0.34.0
python-multipart>=0.0.18

# --- UI Prototype (Phase 4) ---
streamlit>=1.41.0

# --- Testing ---
pytest>=8.3.0
